fixes in .c and make file creation

+++ Projects/Libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: hariskon <hariskon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/24 16:22:51 by hariskon          #+#    #+#             */
-/*   Updated: 2025/05/02 15:12:09 by hariskon         ###   ########.fr       */
+/*   Updated: 2025/05/02 17:14:55 by hariskon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void	*ft_memchr(void *array, int value, size_t size)
 {
-	int				i;
+	size_t				i;
 	unsigned char	p;
 
 	i = 0;
